Added Device Pixel Ratio ( dpr ) to awstats_misc_tracker.js query string
modern monitors have varying device pixel ratios. Small screens such as IPhones use device pixel ratios ( dpr ) of higher than 1. Some have dpr of 2, some 3 and some as high as 5 or 6. But its not only mobile devices that use pixel doubling or trebling etc. The new 4K and 5K monitors use it as well.
Devices that use a dpr of not equal to 1 alter the window.screen width and height in the DOM according to the dpr. So an apple 5K monitor that has a true resolution of 5120x2880 pixels and uses a dpr of 2, returns window.screen width and height of 2560x1440.
It so happens that Apple produce a non retina monitor which has a dpr of 1 and window.screen of 2560x1440. 
Currently, the awstats screen sizes report will add both of these very different monitors to the same count which is technically incorrect.
I'm not perl programmer and don't know how to rectify the code to count correctly and output the DPR in the screen sizes report. However,