Web Page tracking
Added device pixel ratio ( dpr ) to awstats_misc_tracker.js.
This has the effect of showing dpr in your web log files  and is in addition to the already shown screen and window sizes.
It has no effect on the current screen report in awstats but should really be added to that report ( which is not an insignificant task for a perl programmer). However you can see it in your web log files and gives a better indication of device type used to view your pages.